+++ packages/in_app_purchase/example/android/app/src/test/java/io/flutter/plugins/inapppurchase/MethodCallHandlerTest.java
@@ -285,7 +285,6 @@ public class MethodCallHandlerTest {
     verify(mockBillingClient).launchBillingFlow(any(), billingFlowParamsCaptor.capture());
     BillingFlowParams params = billingFlowParamsCaptor.getValue();
     assertEquals(params.getSku(), skuId);
-    assertNull(params.getAccountId());
 
     // Verify we pass the response code to result
     verify(result, never()).error(any(), any(), any());
@@ -312,14 +311,14 @@ public class MethodCallHandlerTest {
   }
 
   @Test
-  public void launchBillingFlow_ok_AccountId() {
+  public void launchBillingFlow_ok_ObfuscatedAccountId() {
     // Fetch the sku details first and query the method call
     String skuId = "foo";
-    String accountId = "account";
+    String account = "account";
     queryForSkus(singletonList(skuId));
     HashMap<String, Object> arguments = new HashMap<>();
     arguments.put("sku", skuId);
-    arguments.put("accountId", accountId);
+    arguments.put("obfuscatedAccountId", account);
     MethodCall launchCall = new MethodCall(LAUNCH_BILLING_FLOW, arguments);
 
     // Launch the billing flow
@@ -337,7 +336,38 @@ public class MethodCallHandlerTest {
     verify(mockBillingClient).launchBillingFlow(any(), billingFlowParamsCaptor.capture());
     BillingFlowParams params = billingFlowParamsCaptor.getValue();
     assertEquals(params.getSku(), skuId);
-    assertEquals(params.getAccountId(), accountId);
+
+    // Verify we pass the response code to result
+    verify(result, never()).error(any(), any(), any());
+    verify(result, times(1)).success(fromBillingResult(billingResult));
+  }
+
+  @Test
+  public void launchBillingFlow_ok_ObfuscatedProfileId() {
+    // Fetch the sku details first and query the method call
+    String skuId = "foo";
+    String profile = "profile";
+    queryForSkus(singletonList(skuId));
+    HashMap<String, Object> arguments = new HashMap<>();
+    arguments.put("sku", skuId);
+    arguments.put("obfuscatedProfileId", profile);
+    MethodCall launchCall = new MethodCall(LAUNCH_BILLING_FLOW, arguments);
+
+    // Launch the billing flow
+    BillingResult billingResult =
+        BillingResult.newBuilder()
+            .setResponseCode(100)
+            .setDebugMessage("dummy debug message")
+            .build();
+    when(mockBillingClient.launchBillingFlow(any(), any())).thenReturn(billingResult);
+    methodChannelHandler.onMethodCall(launchCall, result);
+
+    // Verify we pass the arguments to the billing flow
+    ArgumentCaptor<BillingFlowParams> billingFlowParamsCaptor =
+        ArgumentCaptor.forClass(BillingFlowParams.class);
+    verify(mockBillingClient).launchBillingFlow(any(), billingFlowParamsCaptor.capture());
+    BillingFlowParams params = billingFlowParamsCaptor.getValue();
+    assertEquals(params.getSku(), skuId);
 
     // Verify we pass the response code to result
     verify(result, never()).error(any(), any(), any());
@@ -350,10 +380,8 @@ public class MethodCallHandlerTest {
     MethodCall disconnectCall = new MethodCall(END_CONNECTION, null);
     methodChannelHandler.onMethodCall(disconnectCall, mock(Result.class));
     String skuId = "foo";
-    String accountId = "account";
     HashMap<String, Object> arguments = new HashMap<>();
     arguments.put("sku", skuId);
-    arguments.put("accountId", accountId);
     MethodCall launchCall = new MethodCall(LAUNCH_BILLING_FLOW, arguments);
 
     methodChannelHandler.onMethodCall(launchCall, result);
@@ -368,10 +396,8 @@ public class MethodCallHandlerTest {
     // Try to launch the billing flow for a random sku ID
     establishConnectedBillingClient(null, null);
     String skuId = "foo";
-    String accountId = "account";
     HashMap<String, Object> arguments = new HashMap<>();
     arguments.put("sku", skuId);
-    arguments.put("accountId", accountId);
     MethodCall launchCall = new MethodCall(LAUNCH_BILLING_FLOW, arguments);
 
     methodChannelHandler.onMethodCall(launchCall, result);
@@ -505,7 +531,6 @@ public class MethodCallHandlerTest {
 
     ConsumeParams params =
         ConsumeParams.newBuilder()
-            .setDeveloperPayload("mockPayload")
             .setPurchaseToken("mockToken")
             .build();
 
@@ -539,7 +564,6 @@ public class MethodCallHandlerTest {
 
     AcknowledgePurchaseParams params =
         AcknowledgePurchaseParams.newBuilder()
-            .setDeveloperPayload("mockPayload")
             .setPurchaseToken("mockToken")
             .build();