+++ packages/in_app_purchase/lib/src/billing_client_wrappers/billing_client_wrapper.dart
@@ -147,9 +147,9 @@ class BillingClient {
   /// Attempt to launch the Play Billing Flow for a given [skuDetails].
   ///
   /// The [skuDetails] needs to have already been fetched in a [querySkuDetails]
-  /// call. The [accountId] is an optional hashed string associated with the user
+  /// call. The [obfuscatedAccountId] is an optional hashed string associated with the user
   /// that's unique to your app. It's used by Google to detect unusual behavior.
-  /// Do not pass in a cleartext [accountId], use your developer ID, or use the
+  /// Do not pass in a cleartext [obfuscatedAccountId], use your developer ID, or use the
   /// user's Google ID for this field.
   ///
   /// Calling this attemps to show the Google Play purchase UI. The user is free
@@ -165,14 +165,17 @@ class BillingClient {
   /// [`BillingFlowParams`](https://developer.android.com/reference/com/android/billingclient/api/BillingFlowParams)
   /// instance by [setting the given
   /// skuDetails](https://developer.android.com/reference/com/android/billingclient/api/BillingFlowParams.Builder.html#setskudetails)
+  /// [the given
+  /// obfuscatedAccountId](https://developer.android.com/reference/com/android/billingclient/api/BillingFlowParams.Builder.html#setObfuscatedAccountId(java.lang.String)).
   /// and [the given
-  /// accountId](https://developer.android.com/reference/com/android/billingclient/api/BillingFlowParams.Builder.html#setAccountId(java.lang.String)).
+  /// obfuscatedProfileId](https://developer.android.com/reference/com/android/billingclient/api/BillingFlowParams.Builder.html#setObfuscatedProfileId(java.lang.String)).
   Future<BillingResultWrapper> launchBillingFlow(
-      {@required String sku, String accountId}) async {
+      {@required String sku, String obfuscatedAccountId, String obfuscatedProfileId}) async {
     assert(sku != null);
     final Map<String, dynamic> arguments = <String, dynamic>{
       'sku': sku,
-      'accountId': accountId,
+      'obfuscatedAccountId': obfuscatedAccountId,
+      'obfuscatedProfileId': obfuscatedProfileId,
     };
     return BillingResultWrapper.fromJson(
         await channel.invokeMapMethod<String, dynamic>(
@@ -226,18 +229,15 @@ class BillingClient {
   /// Consumption is done asynchronously. The method returns a Future containing a [BillingResultWrapper].
   ///
   /// The `purchaseToken` must not be null.
-  /// The `developerPayload` is the developer data associated with the purchase to be consumed, it defaults to null.
   ///
   /// This wraps [`BillingClient#consumeAsync(String, ConsumeResponseListener)`](https://developer.android.com/reference/com/android/billingclient/api/BillingClient.html#consumeAsync(java.lang.String,%20com.android.billingclient.api.ConsumeResponseListener))
-  Future<BillingResultWrapper> consumeAsync(String purchaseToken,
-      {String developerPayload}) async {
+  Future<BillingResultWrapper> consumeAsync(String purchaseToken) async {
     assert(purchaseToken != null);
     return BillingResultWrapper.fromJson(await channel
         .invokeMapMethod<String, dynamic>(
             'BillingClient#consumeAsync(String, ConsumeResponseListener)',
             <String, String>{
           'purchaseToken': purchaseToken,
-          'developerPayload': developerPayload,
         }));
   }
 
@@ -258,18 +258,15 @@ class BillingClient {
   /// details.
   ///
   /// The `purchaseToken` must not be null.
-  /// The `developerPayload` is the developer data associated with the purchase to be consumed, it defaults to null.
   ///
   /// This wraps [`BillingClient#acknowledgePurchase(String, AcknowledgePurchaseResponseListener)`](https://developer.android.com/reference/com/android/billingclient/api/BillingClient.html#acknowledgePurchase(com.android.billingclient.api.AcknowledgePurchaseParams,%20com.android.billingclient.api.AcknowledgePurchaseResponseListener))
-  Future<BillingResultWrapper> acknowledgePurchase(String purchaseToken,
-      {String developerPayload}) async {
+  Future<BillingResultWrapper> acknowledgePurchase(String purchaseToken) async {
     assert(purchaseToken != null);
     return BillingResultWrapper.fromJson(await channel.invokeMapMethod<String,
             dynamic>(
         'BillingClient#(AcknowledgePurchaseParams params, (AcknowledgePurchaseParams, AcknowledgePurchaseResponseListener)',
         <String, String>{
           'purchaseToken': purchaseToken,
-          'developerPayload': developerPayload,
         }));
   }