+++ packages/in_app_purchase/lib/src/billing_client_wrappers/purchase_wrapper.g.dart
@@ -20,6 +20,8 @@ PurchaseWrapper _$PurchaseWrapperFromJson(Map json) {
     isAcknowledged: json['isAcknowledged'] as bool,
     purchaseState:
         const PurchaseStateConverter().fromJson(json['purchaseState'] as int),
+    obfuscatedAccountId: json['obfuscatedAccountId'] as String,
+    obfuscatedProfileId: json['obfuscatedProfileId'] as String,
   );
 }
 
@@ -33,6 +35,8 @@ Map<String, dynamic> _$PurchaseWrapperToJson(PurchaseWrapper instance) =>
       'sku': instance.sku,
       'isAutoRenewing': instance.isAutoRenewing,
       'originalJson': instance.originalJson,
+      'obfuscatedAccountId': instance.obfuscatedAccountId,
+      'obfuscatedProfileId': instance.obfuscatedProfileId,
       'developerPayload': instance.developerPayload,
       'isAcknowledged': instance.isAcknowledged,
       'purchaseState':